Add the Lumi TFGen command line
This change introduces the Lumi TFGen command line tool.

This tool generates the Lumi package metadata for a collection of Terraform
providers.  This is done through static linking due to the way that Terraform
hides all schema knowledge behind the dynamic plugin interface in a way that
resists dynamic inspection.

A single dynamic plugin may be used to drive any number of such Lumi
packages.  Although we must generate the metadata N times, at least we need
not have N runtime plugin executables to support them.  This is just so that
we can get some compile-time validation against the Terraform schema.

Each such module has a flat list of exports, each corresponding to a resource,
although by the time we are done we must implement some additional sub-module
structure.  This will ensure, for example, that there aren't 300 resources
at the top-level AWS module, and that instead we have EC2, S3, etc. sub-modules.

In addition to this, there is a single idiomatic Lumi config module.  This
contains all of the Terraform provider variables so that they can be configured
in the usual Lumi way.  For example, the AWS region, auth tokens, and so on.

There are a number of loose ends in here, mostly marked with
TODOs/FIXMEs/etc., and we have not yet implemented name mangling (so the
names are still ugly_underscore_names).  Nevertheless, with this checkin we
can begin driving the CRUD operations through the single plugin provider.