Implement an initial tfbridge resource provider
This commit includes a rudimentary Terraform bridge resource provider.
It is capable of Lumi CRUD operations against any Terraform module and,
modulo the need for test cases, fixing some TODOs/etc., and shaking out
some bugs, is surprisingly complete for the small amount of code in here.

The basic approach here is to load the Terraform plugins dynamically.
This may initially seem funny -- since Lumi loads this provider as a
plugin, only for it to turn around and load another plugin -- but comes
with some benefits.  For one, we can operate "at an arm's length" using a
real RPC interface meant for dynamic loading scenarios just like this one.
This gives us a certain amount of loose coupling that, we hope, will
insulate us from changes to the internal guts of Terraform.  Another is
that we get to write a single implementation of the bridge that works
against any possible Terraform plugin in the future, at least those that
conform to the current version of the interface.

A few specific notes:

* We configure the Terraform plugin with all variables in the idiomatic
  "config" module that the Lumi TFGEN tool generates.  This is done simply
  by enumerating the heap state for each variable and poking it into the
  plugin via the Configure RPC call.  This is done at potentially
  inopportune times -- basically when we load the plugin -- and so we will
  need to think harder about configuration lifetime as a next step.

  This means for example that you can use the standard Lumi config facility

    $ lumi config tf-aws:config:region us-east-2

  and that this flows naturally to Terraform provider configuration state.

* Check performs all Terraform property validation, as you would expect.

* Name simply uses the given name to auto-name things in the URN, since all
  TFGEN resources are named resources, much like we do in other Lumi providers.

* Create, Update, and Delete all utilize the Terraform plugin Apply routine
  to create states and diffs that carry out the intended behavior.  It's
  a little funky that we must generalize from Create, Update, and Delete to
  a very general-purpose Apply operation, only for it to internally re-
  materialize the Create, Update, and Delete-nature of the operation within
  Terraform's pkg/helper/schema/ module, but hey, that's the way it is.
  Besides, there could conceivably be Terraform providers not using schema.

* Get just refreshes an empty state using the Terraform resource ID.

* InspectChange relies on the ability to invoke the Diff routine with a
  proposed change, to perform resource-specific diffing, such that we can
  then inspect the RequiresNew fields on the resulting object to determine
  whether replacement will be required.

Given this, our basic web server example works with the Terraform bridge!