Switch from Godep to Go's new Dep tool
This changes over to Go's new Dep tool.  It is still in Alpha, so
I'm proceeding with caution and using it in the terraform-bridge
repo as a dry run.  The Godep tool is showing age and keeps resolving
diamonds to the wrong version, resulting in unbuildable code.  Dep
seems to do a better job at managing this.