Generate config as a sub-directory module
This changes the emission of config modules to be underneath a
directory, rather than simply a file named config.ts.  The reason
for this is that we're about to add support for "overlays", which
permits augmentation of the auto-generated code with helper functions,
modules, and additional typing.  These sometimes go into the config
module itself and we want to avoid needing to textually merge source.