Add the notion of overlays
This change introduces overlays.  An overlay is simply a manual
augmentation of an otherwise purely code-generated package.  This can
be useful if you want to add helper functions, or even entire modules,
alongside the generated code.  It can also be used to add gradual types
like enums that can be used to replace existing Terraform weakly typed
variables like strings (e.g., regions, instance types, and so on).