Use a capability for Instance->SecurityGroup references
This changes the API we project to accept a SecurityGroup reference,
rather than just a name, for Instance's securityGroups property.  This
mirrors the way our original Lumi package did this and preserves the
nice programming model of doing this

    let group = new aws.ec2.SecurityGroup(...);
    let server = new aws.ec2.Instance(..., {
        ...,
        securityGroups: [ group ],
    });

instead of needing to do this, which is far less discoverable

    let group = new aws.ec2.SecurityGroup(...);
    let server = new aws.ec2.Instance(..., {
        ...,
        securityGroups: [ group.name ],
    });

This resolves pulumi/terraform-bridge#2.