Adopt a different strategy for names and collisions
This change adopts a completely different strategy for names and
collisions between Terraform and Lumi names.

First thing's first, any collisions must be explicitly addressed
through an overlay.  Otherwise, we end up doing the correct thing
1/2 of the time and the incorrect thing the other 1/2 of the time.
As a result, an error is issued at tfgen time.  To make chasing
down many errors a little more palatable, we report many errors
at once at the command line, rather than stopping at the first.

Second, we permit overlays to include three new things:

1) A default value.
2) A directive to take the value of another field as the default.
3) A directive to transform such a field taken as the default.

The way we use this for names is that, on the backend, we still
have the distinct Terraform names (which map to the underlying
raw resource names on the cloud provider).  So, we can simply
propagate the Lumi "name" property, using (2) above.

This alone simply adds essentially an aliased input property.
But we want to go further than this, since naming the backend
resources like this creates the opportunity for conflicts.  As
a result, we also permit a custom function that will transform
the value in some way.  We can use this to append a hash like
our current resource providers generally do.  We will *not* do
this if the developer specifies this name explicitly, which can
be used in circumstances where s/he knows and understands the
implications of naming like this (multi-instantiation woes).

For example, the resulting overlay looks something like:

    "aws_sqs_queue": {
        Fields: map[string]SchemaInfo{
            "name": {
                Name: "queueName",
                Default: {
                    From: "name",
                    FromTransform: func(v interface{})interface{}{
                        return resource.NewUniqueHex(
                            v.(string)+"-", 255, -1)
                    },
                },
            },
        },
    },

To make this common pattern easier to express, this change also
includes an autoName function.  The above example becomes simply:

    "aws_sqs_queue": {
        Fields: map[string]SchemaInfo{
            "name": autoName("queueName", ..., 255),
        },
    },