Auto-name AWS/Azure/GCP resources
This change applies auto-naming to all AWS, Azure, and GCP
resources.  Surely we have gone too far with this, but it's an
expedient way to get the auto-naming deployed to all of the
resources that need it.  Out of the 500-or-so AWS resources,
200-or-so have names; my guess is that out of those, only 150
should have this change.  (Any that CloudFormation warns you
not to explicitly name due to inability to replace resources.)

In summary:

* Add an autoName routine that injects a resource auto-name.

* Fix the reverse lookup logic, in addition to applying
  defaults appropriately using the right Lumi and Terraform names.

* Add some additional logging that came in handy.

* Fix up some optional logic so it accounts for default values.

* Auto-enumerate all resources for the providers, and perform
  auto-name translation in any case where there's a conflict.

It's not yet clear that we want to apply this to Azure and GCP,
but it's a quick way to get through the name conflicts.  We will
need to audit this more deeply as we get to those providers.