Introduce composite keys
Lumi prefers to have singular identifiers (primary keys, if you will)
for its resources.  This simplifies many tasks, because we have a single
string that we can serialize or materialize to uniquely identify objects.

Terraform, on the other hand, is a bit "looser" when it comes to identity,
despite it having an ID field on resources.  This is really unfortunate
for a number of reasons, like the difficulty it introduces in Lumi's ability
to have a clean "get" API to lookup existing object state.  Most resources
do in fact use nothing but the ID, but there are others that rely more on
"value equality" than "identity/reference equality", in a sense.

One option would be to cave and give up on a singular ID, and change our
Get and Delete methods to take whole bags of state, like Terraform does.  If
we can model the need to have additional fields as a "filter" of sorts, this
might not be too terrible, especially as introducing queries for resources will
probably take us here.  But it's a rather big divergence from our current model
and would complicate the simple lookup scenarios, with strange failure modes.

Instead of doing that, this change introduces the notion of composite keys.
The idea is that for simple resources with singular IDs, they can continue
to use singular IDs just like today.  This is the majority case.  For others,
however, a resource may declare that it uses multiply fields to form its
identity.  The bridge provider then manages the creation and parsing of them.

A composite key is simply an ID, as usual, followed by a ";" and any number of
key/value pairs in the form "key=value", delimited by ";"s.

This is mostly an implementation detail of the provider, except that a user
may notice IDs that look like key-value strings.  For instance, a an IAM
RolePolicyAttachment in one of my test cases has the following ID:

mylambda-role-66bfbcf5eec8a3f5045f0dd13f7c41f297499102-000485c28e9f60d33d8c084801;
    policyArn=arn:aws:iam::aws:policy/AWSLambdaFullAccess;
    role=mylambda-role-66bfbcf5eec8a3f5045f0dd13f7c41f297499102

Although messy, it does reflect reality: in other words, we need the
policyArn and role fields in order to Get or Delete the resource.