Add the concept of ID fields
Sometimes a provider will sloppily use a field whose value is identical
to the ID, rather than using the resource ID.  In such cases, arguably
the provider should have and could have used the ID, but the temptation
of having the bag of state available means this will happen.

For example, instead of

    query(d.ID())

the provider will do something like

    query(d.Get("name"))

To remedy this, we could use composite keys, but that's overkill.  We'd
end up with silliness like "foo1234abcd;name=foo1234abcd".  Instead, we
will permit marking certain fields as IDFields in the overlays.  So,
in the above example, we have []string{"name"}, and the Get/Delete functions
will patch up the Terraform state as required to perform the query.