Be more precise about inputs/outputs
This changes the way we generate properties slightly.  Terraform has
three flags any property can hold: Required, Optional, and Computed.

Required has the obvious meaning: users must supply the value.  Optional,
similarly, means users may, but aren't required to, supply a value.

Something not necessarily obvious, however, is that there are two other
interesting cases.  First, something may be Computed but neither Required
nor Optional.  This means the property is *never* supplied by the user.
Instead, it is always an output property.  The next case is that something
is Optional *and* Computed, to mean that a user may supply a value but,
in the event that it is omitted, a value will be computed by the provider.

This is now handled more precisely so users get a nice typing experience:

* Required: on the resource and args objects.
* Optional: on the resource and args objects, but marked optional (?).
* Computed (+Optional): optional on the arg object (?), always on the resource.
* Computed (not Required, not Optional): on the resource object, not args.

This fixes up some annoying non-null errors you'd get trying to access
legitimate output properties, like ARNs and whatnot.