Populate Terraform defaults
This change correctly populates the Terraform schema default values,
and fixes pulumi/terraform-bridge#4.  Somewhat bizarrely, this only
occurs if we invoke Diff in advance of Create.

The reason is fairly roundabout, explained in a comment, however here
it is again in a nutshell, for posterity.

The pkg/helper/schema/ package in Terraform is what provides various
services for resource schemas.  The default values are stored on such
schemas.  The question then becomes under what circumstances will
Terraform discover that a field value is missing and then consult
the schema to perform default initialization?

One would expect this happens for all reads of all schema fields,
as I originally did.  But alas, this is not true.  There are many
field reader abstractions: config, diff, map, and multi, specifically.
These get chained and the resource state roots a multi reader that
will consult a series of other readers in a fixed order.  The only
reader that respects the schema's default value, however, is config!
This means that only if config gets chained in does the field reader
actually discover that fields are missing and consult the defaults.

This is where things go awry.  The config reader is only chained in
certain circumstances.  ValidateResource and Diff, specifically.
Notably, it is *not* chained in for the Apply operation, which is what
is called for CxUD operations.  And although we do call ValidateResource,
we weren't calling Diff, since Lumi is handling diffing of the object
graphs.  However, ValidateResource doesn't communicate back the
resulting default state.  But Diff does.  And there's our solution.

This change adds a call to Diff just before Create, which has the
effect of populating any missing default properties.  We simply pass
that result in to the Create operation, alongside the existing state,
and voila, the right thing happens.