Eliminate custom keys; leverage Delete state
Terraform providers often require arbitrary state from the object
in order to get the job done.  This is a fundamental mismatch with
Lumi's model of deletes only requiring the ID field.  In fact, many
Terraform providers operate only in terms of the ID, however, many
do not.  Instead of chasing down the long tail of these and manually
working around them with custom key schemas -- a task whose end is
neither in sight nor guaranteed to last as we continue to take updates
to the providers -- we will simply pass the object state into the
Delete API and let it leverage as much of it as it wishes.

This fixes pulumi/terraform-bridge#8.