Parse package version from Dep lockfile, when present
Now that we are starting to vendor some of the Terraform dependencies,
we cannot simply rely on the presence of GOPATH Git repos for them.
This is currently required to fetch the commit/tag information; it is
purely informational, for the pack/VERSION file, but this will be
important for debugging purposes.  Instead of demanding that there be
a Git repo, we will prefer to look for a Dep lockfile, Gopkg.lock,
and use the version listed within; if that doesn't exist, we will
fall back to the existing Git repo logic, as it's conceivable we'll
have some packages that don't vendor (at least to start with).