Expand flatmaps for default values
The way we relayed default values back from Terraform to Lumi
wasn't correct.  We used the unexpanded values, which meant two
things: First, every value was a string, when we of course are
expecting bools, numbers, strings, arrays, and so on; Second,
arrays, maps, and sets came back in their flat form, which meant
lots of garbage keys like `p.#`, `p.2401560585.x`, `p.2401560585.y`,
`p.%`, and so on.  This made me think we weren't marshaling sets
correctly (pulumi/pulumi-terraform#19), but it turns out we are
(diffing aside).  Instead, we need to expand the flatmap before
deciding on default values, and then use the expanded values.