Invoke terraform provider in memory
We were previously spinning up a seperate process with the terraform provider and speaking to it through an RPC interface.

However, we actually already had the implementation of the interface we need to talk to in memory in the tfbridge process already.

So just provide that instance instead of a stub that talks over an RPC interface.

This fixes #15, and should greatly simplify the Travis and Dockerfile environments for running Pulumi programs, as they no longer need an installed terraform-provider-aws binary.