Rework file asset translation.
Note: for the purposes of this discussion, archives will be treated as
assets, as their differences are not particularly meaningful.

This commit solves one fundamental issue with file asset translation and
addresses an inefficiency.

First, translating an asset to a file currently requires that the
contents of the asset are available. This is not the case when
performing a diff, update, or delete: in the former two cases, the
contents of old assets are not available, and in the latter, the
contents of the deleted assets are not available.

Second, we currently write a new file each time we translate an asset to
a file. For archives especially, this can require a prohibitive amount
of CPU, IO, and storage resources.

These changes can both be addressed by making use of the fact that
assets are identifiable by their hash.

For assets with a hash, we can derive the translation file's name from
the hash and make a best effort to reuse an already-translated asset
with the same hash. If the asset has a hash but no contents, we can
simply return the derived file name. This has the additional benefit of
preventing false asset diffs in the case that we are diffing a resource
that has changes to non-asset properties.

For assets without a hash, we fall back to the existing strategy of
creating a new temporary file to hold the translated asset and returning
its path. If such an asset has no contents, we return the empty string,
which is guaranteed to differ from all hashed or non-empty assets.

This is the other half of the fix for pulumi/pulumi-cloud#158. Note that
these changes depend on pulumi/pulumi#548.