Manually apply TF defaults
Prior to this change, we applied Terraform schema defaults in a
massively convoluted way.  Because Terraform applies these deep
down in the bowels of its schema/ConfigFieldReader, we had to
drive a Diff with InstanceState and ResourceConfig, even though
strictly a diff like this isn't required.  That has the consequence
of diffing the "old" state without the defaults, which tickles the
ConfigFieldReader, which causes it to inject default values into
the "new" state which is returned by the Diff.  We then proceeded
to return just the different state back to Pulumi as defaults.

This caused a subtle bug.  In particular, we ended up treating
"policies" as a default property, even though it was included in
the inputs.  At that point, switching to a "policy" on an IAM
InstancePolicy fails, because you cannot set both properties.
This is in turn caused by behavior in the innards of Diff which
detects when the diff RequiresNew, and cooks up a diff that
makes it look like all new state is new (thanks to "path" having
a default value while also being RequiresNew), which caused us
to think that *all* properties were defaults.  (Ugh.)

Not to mention a lot of this field reading code is some of the
most confusing code I've ever seen, with ample TODOs and zero
documentation.

My fix avoids all of it.  Instead of even going down these paths,
we instead just apply the DefaultValue() directly from the TF
schema object.  After doing that to produce the TF inputs, which
we need in order to call Validate() anyway, we then produce our
own diff by comparing the inputs and this resulting set of
properties, and only return as defaults what we believed is new.

This fixes pulumi/pulumi#586 and I'm working on some regression
tests before submitting.