Assert that TF state != nil on Create; lock to v0.10.8
This change asserts that TF state != nil on Create; it is required
to obtain the resulting ID of the created resource.

It appears TF has a bug where this will ocassionally and
indeterminitely return nil.  We suspect this is due to the line
https://github.com/hashicorp/terraform/blob/27aa4cb6bf99b4edcbf473e720c322cf6cffe490/helper/schema/resource_data.go#L318
although we aren't entirely sure.  (That line is suspicious and
there has been some recent activity in and around the code that
deals with such errors.)  To confirm our suspicion, to my
knowledge we've only seen this in repos that have locked to v0.11.0
of Terraform, which was released within the past week.

So, for now, I am also going to lock to v0.10.8 in hopes that
this will clear things up while we dig deeper into
pulumi/pulumi-terraform#57.