Return all provider inputs from `Check`.
These changes update to the latest `pulumi/pulumi` and adopt the new
semantics for `Check`. In the new model, it is the provider's
responsibility given any old inputs and the new inputs for a resource to
merge any appropriate defaults from the old inputs into the new inputs,
then return the full set of provider inputs. The TF provider performs
this by preferring any defaults that were present in the old inputs to
those produced by the Pulumi overlay or Terraform schema.