Map marshaled assets back to their original value. (#75)
As a rule, the provider inputs returned by a call to `Check` should
preserve the original representation of the properties as present in the
program inputs. While this was the case for most property types, it was
not the case for assets/archives, which were being returned in their
marshaled form. This change adds support for tracking which inputs were
assets in a side table that can be shared between
MakeTerraform{Inputs,Outputs} in order to perform the inverse mapping.

This is the primary fix for pulumi/pulumi#645.