Permit alternative types (#85)
This change adds the ability to create union types for Terraform inputs.
This is required because now we often flow values as configuration across
stacks -- for multi-stack linking -- and so, for instance, we won't
necessarily have a strongly typed "capability" reference to supply.  In
the future, we will have ways to look up resources by name, etc.
(see https://github.com/pulumi/pulumi/issues/83), but this probably always
needs to exist as a workaround anyway.