Permit nil diffs in Update.
It is very possible for us to get encounter a nil diff in Update:
custom diffing behavior, etc. can cause textual/structural changes not
to be semantic changes. A better solution would be to change the
result of Diff to indicate no change, but that is a slightly riskier
change that we'd rather not take at the current moment.

Fixes #94.