Restructure TF code-gen to support multiple languages
The prior structure of the TF code-gen logic was very much *not*
amenable to multiple language support.  Type/JavaScript language-isms
ran throughout, without any of the needed virtualization.

This change breaks apart this code, and creates a langGenerator
interface that may be implemented for each language.  All concerns
that are language-specific -- like type names, how to lay out modules,
and so on -- are on the other side of the langGenerator interface,
while the generator itself still understands how to build up the
logical structure of a package (types, config, and so on), and drive
the langGenerator to emit the right code accordingly.

The only real implementation, of course, is Node.js/TypeScript.