Move type logic into language providers
The prior attempt at creating the right boundary between the generator
and language-specific logic attempted to gather type information eagerly,
associating it with the various variable, type, and function objects
subsequently used during code emission.  Although technically this could
probably work, it's problematic, because the decision about what type
information to emit -- if any! -- often depends on some contextual
information only known by the language generator at a given moment in time.

For instance, whether to generate a type as `Foo | undefined`, or just
`Foo`, in TypeScript requires knowledge of whether that type is intended to
be used in a property slot already annotated with a `?`, as a generic type
argument, perhaps deeply embedded in another structure, whether the target
type or function being instantiated already modifies the return type to have
a `| undefined` component, and so on.  In short, it's tricky.

The easiest thing to do is just to leave it entirely up to the language
generator to decide how and when to create these types and flags.