Update pulumi{-terraform} and switch to DeleteBeforeReplace
Until we are auto-generating `metadata.name` with a random suffix by
default, we must mark all Kubernetes resources as delete-before-replace.
Without this, any change which forces a replace will re-use the same resource
name, which will fail.

We will be able to reconsider this default once pulumi/pulumi-kubernetes#10
is fixed.