Implement `kubeProvider#Check`
This commit will introduce `kubeProvider#Check`, which takes a Kubernetes
API object, and verifies that it is in compliance with an OpenAPI schema
that is retrieved from a live API server. This schema represents the
"capabilities" of an API server (e.g., whether flags turned RBAC on or
off) as well as CRDs (assuming CRD schemas are provided for relevant
resources), so it captures much of the variation in real-world running
API servers.