Introduce k8s dependencies at known-good pinned versions
The ksonnet project uses a known-good (-ish!) set of pinned versions of
k8s client dependencies. There are a number of practical reasons for
this; most of them have to do with the fact that client-go uses some
"controversial" techniques for specifying its dependencies, which will
cause `dep` to not rope in the correct versions of some of its
dependencies out of the box.

This commit will save us the time of re-figuring-out all these
dependency problems by introducing these exact dependencies to our
Kubernetes provider.