Introduce blocking `Create` for v1 Kubernetes objects
In other Kubernetes command orchestration (e.g., the Terraform
Kubernetes provider), a `Create` call will issue a request to create
some set of resources, and then attempt to block until the object is
fully initialized. For example, a replication controller deploying 1k
containers will not be instant, and if it is important for these to be
started before the next operation, we will need mechanisms to wait.

This commit will introduce blocking-with-timeout semantics for `Create`
operations for the Kubernetes objects in "stable" (i.e., non-alpha and
non-beta) APIs. This logic is inspired to a great extend by the
Terraform Kubernetes provider, but this is really meant to be the
starting point, as we expand to support (e.g.) alpha and beta APIs as
well.

To be more specific, this commit will introduce:

- The `await` package, from which users are meant to interact with our
  API for awaiting object initialization. See, for example, the code in
  our Kubernetes provider responsible for this await:

    initialized, err := await.Creation(k.pool, k.client, obj)

  Underlying this is a collection of polling/await/retry primitives, and
  several polling strategies for the "stable" Kubernetes API objects.
- The `client` package, which now contains all the client bits that were
  located in the `provider` package. This is to make them accessible
  both to the polling logic in `await` and the provider logic in
  `provider`.

The current roadmap for this feature is:

- [x] Add `await` logic for the `Create` operation for the "stable"
      Kubernetes API.
- [ ] Comprehensive unit and integration tests.
- [ ] Similar logic for `Delete` and `Update`.