Implement blocking `Delete` operation
See the previous commit for a detailed description of how blocking
operations are to be implemented in our Kubernetes provider.
Essentially, this commit will introduce the ability to wait for
resources to gracefully terminate (e.g., scale down the replicas).

Similarly to the last commit, the main entry point to this await logic
is the function:

  err := await.Deletion(k.pool, k.client, gvk, namespace, name)

The essential nature of this call is to poll until we are sure that all
resources have been gracefully deleted, or there is an error.