Implement blocking `Update` operation
See the previous commit for a detailed description of how blocking
operations are to be implemented in our Kubernetes provider.
Essentially, this commit will introduce the ability to wait for
resources to gracefully update (e.g., scale up replicas).

Note that this commit will not change the behavior of update, which is
to say that this commit still implements three-way JSON merge patch
semantics.  (Incidentally, in the process of writing this commit, we
learned that the Terraform Kubernetes provider actually uses this
strategy exclusively.)

Similarly to the last commit, the main entry point to this await logic
is the function:

  _, err := await.Update(k.pool, k.client, oldObj, newObj)

The essential nature of this call is to poll until we are sure that all
resources have been gracefully deleted, or there is an error.