FIX: Differences are not reported during `preview`
CAUSE: The current implementation of `kubeProvider#Diff` does not
correctly report if there are changes between two version of an object.
This has gone unnoticed because (I suspect) `DiffResponse.Changes` has
been ignored in the implementation of the Pulumi update protocol until
the most recent release, v0.12.0.

FIX: Report that two version of a Kubernetes object are different if
there is any diff at all.