FIX: Every `update` results in a hard replacement of all objects
CAUSE: The output values of `kubeProvider#Create` and
`kubeProvider#Update` are supposed to be the "live" version of the
Kubernetes API objects after they are fully initialized. A
misunderstanding of these semenatics led us to report computed values as
a list of JSON path-based identified properties that were computed by
the server. In v0.12.0, this causes every `update` to hard-replace all
Kubernetes resources in the stack, because it can't find `namespace` or
`name` in the output object. Since replacement (at this point) is
determined by whether the old and new values of these properties are
equal, and since these properties will never be equal (since all objects
always have names and namespaces), Pulumi concludes that we must always
replace every object. I suspect this went unnoticed because these
computed properties were not really used in v0.11 in the same way they
are now.

FIX: This commit will simply retrieve the live version of the API object
after initialization, and report that as the output value.