Correctly identify when `update` results in replacement
The Pulumi provider gRPC interface currently does not give `Diff` both
the last version of the submitted application and the current version.
In other cloud providers, this is not a big deal, but in Kubernetes,
this means we can't decide whether an update will result in a
"create-and-delete" replacement, because this decision depends on
whether the _user_ has changed a forcenew property between the last time
they submitted and the current time.

Until this is implemented, we need a work-around. This commit will
introduce a temporary solution, making the output of `Create` and
`Update` return both the inputs and the computed values, so that they
are propagated to `Diff`.