Work-around: minikube fails to report version correctly
Some versions of minikube, when asked for Kubernetes version, will
report a blank string. This will cause us to fail when trying to delete
resources, because we can't parse the blank string, which means we can't
work around the quirks of various Kubernetes versions.

This commit will introduce two fallbacks: first, if the version fails to
parse, we'll try to parse the git version reported by Kubernetes.
Second, if that fails, then we take a wild guess (i.e., v1.9).