Add input and (naive) output Kubernetes API types
Implicit in the Kubernetes API is a notion of types that are optional
when they're provided as input to the API server, and types that are
optional when they're given to the user (e.g., on a GET request). So
far our type generation for TypeScript does not distinguish between
these two things.

This commit will introduce distinct input and output types for each API
object. The input types follow the guidance of the OpenAPI spec, which
indicates whether a field is optional. The output types, on the other
hand, are naive, marking _no_ fields as optional. This is a pragmatic
half-measure until we can assemble a list of which properties actually
are optional.