Simplifying watch code
The watch/await code we use to wait for all resources to be fully
initialized before we continue is heavily informed from the polling code
in Terraform. But, this code is quite a bit more complex than what we
need for our purposes. So, this commit will introduce a much simplified
package that does only what we need it to do. We cut the code in half,
even after adding tests.

In an ideal world, this would depend on the Kubernetes `Watch` API, but
in this case we can't filter on name, and:

* It's not clear what the impact would be on large clusters relative to
  simple polling, particularly for a resource type like pods, where
  there could be a very large number cluster-wide.
* There don't seem to be any serious downsides to polling, and it is
  somewhat simpler to implement.

Fixes #44.