FIX: Default deployment await logic
The await logic for deployments currently keys off `.status.replicas`.
But, this field actually just lists the number of replicas that kubelets
have reported have started. To be considered available, they still have
to be stably available for some (customizable) minimum number of
seconds. This means that our await logic (which is meant to block until
this operation is complete) will nearly always return before _any_ of
the pods are ready.

In this commit, we transition away from this, instead waiting until
`.status.availableReplicas` is equal to `.spec.replicas`.

Relatedly, our delete logic keys off of the status replica fields, but
the only reliable way to know whether all replicas are destroyed is to
check whether the deployment has been deleted from the API server.
Hence, we will replace the delete logic, too.