FIX: checkpoint/update semantics
The current Kubernetes provider is based on some fundamental
misunderstandings on when in request and response objects the field
`.Olds` refers to old _state_ vs. old _inputs_. This commit will correct
them and express these correction in comments. More context follows

Kuberenetes requires 3 API objects to create an `Update` patch:

  1. The current resource supplied to us from the user.
  2. The last version of the resource supplied from the user.
  3. The live version of the resource as it exists in Kubernetes.

The reasons for this are covered fairly well in the documentation for
`kubeProvider#Update`. For the purposes of this patch, what's important
to understand is that the `Update` method does not actually supply the
last version of the resource supplied by the user; it only supplies the
current version and the live version as it exists in Kubernetes.

We plumb this through the gRPC provider by having `Create` and `Update`
return a "checkpoint object", of the form {live: {...}, input: {...}},
which contains both. This is then unpacked/repacked in relevant methods.

Unfortunately, during implementation there were some misunderstandings
about when in request and response objects the field `.Olds` refers to
old _state_ or old _inputs_, and hence, we are unpacking/repacking them
in incorrect places.